requirements.txt
grpcio==1.71.0
grpcio-tools==1.71.0
protobuf~=5.29.4

rasterio~=1.4.3
geopandas
shapely
pyproj
geojson~=3.2.0

elevation

numpy>=1.21,<2.3
matplotlib~=3.10.1
scikit-image
opencv-python
Pillow

requests
tensorflow
huggingface_hub
transformers